Script for auto download and installation of Dyna
This bash script is formulated for automatic download and installation of LS-Dyna in simple Unix (here, Ubuntu) environment. Though, the LS-Dyna release version 11.2.1 is used here, it can be extended to include any of the versions available in the LSTC's ftp server. The license declared is a simple network license supposed to be only working inside the campus network of UNR. Users of the script can simply replace the UNR network license with their own and can proceed.